+++ src/search/goal_count_heuristic.cc
@@ -21,7 +21,7 @@ void GoalCountHeuristic::initialize() {
 
 int GoalCountHeuristic::compute_heuristic(const State &state) {
     int unsatisfied_goal_count = 0;
-    for (int i = 0; i < g_goal.size(); ++i) {
+    for (size_t i = 0; i < g_goal.size(); ++i) {
         int var = g_goal[i].first, value = g_goal[i].second;
         if (state[var] != value)
             ++unsatisfied_goal_count;
