+++ src/search/hm_heuristic.cc
@@ -66,7 +66,7 @@ void HMHeuristic::update_hm_table() {
         ++round;
         was_updated = false;
 
-        for (int op_id = 0; op_id < g_operators.size(); ++op_id) {
+        for (size_t op_id = 0; op_id < g_operators.size(); ++op_id) {
             const Operator &op = g_operators[op_id];
             tuple pre;
             get_operator_pre(op, pre);
@@ -77,7 +77,7 @@ void HMHeuristic::update_hm_table() {
                 vector<tuple> partial_eff;
                 get_operator_eff(op, eff);
                 generate_all_partial_tuple(eff, partial_eff);
-                for (int i = 0; i < partial_eff.size(); ++i) {
+                for (size_t i = 0; i < partial_eff.size(); ++i) {
                     update_hm_entry(partial_eff[i], c1 + get_adjusted_cost(op));
 
                     if (partial_eff[i].size() < m) {
@@ -95,7 +95,7 @@ void HMHeuristic::extend_tuple(tuple &t, const Operator &op) {
         pair<tuple, int> hm_ent = *it;
         tuple &entry = hm_ent.first;
         bool contradict = false;
-        for (int i = 0; i < entry.size(); ++i) {
+        for (size_t i = 0; i < entry.size(); ++i) {
             if (contradict_effect_of(op, entry[i].first, entry[i].second)) {
                 contradict = true;
                 break;
@@ -106,7 +106,7 @@ void HMHeuristic::extend_tuple(tuple &t, const Operator &op) {
             get_operator_pre(op, pre);
 
             tuple others;
-            for (int i = 0; i < entry.size(); ++i) {
+            for (size_t i = 0; i < entry.size(); ++i) {
                 pair<int, int> fact = entry[i];
                 if (find(t.begin(), t.end(), fact) == t.end()) {
                     others.push_back(fact);
@@ -121,7 +121,7 @@ void HMHeuristic::extend_tuple(tuple &t, const Operator &op) {
 
             set<int> vars;
             bool is_valid = true;
-            for (int i = 0; i < pre.size(); ++i) {
+            for (size_t i = 0; i < pre.size(); ++i) {
                 if (vars.count(pre[i].first) != 0) {
                     is_valid = false;
                     break;
@@ -143,7 +143,7 @@ int HMHeuristic::eval(tuple &t) {
     vector<tuple> partial;
     generate_all_partial_tuple(t, partial);
     int max = 0;
-    for (int i = 0; i < partial.size(); ++i) {
+    for (size_t i = 0; i < partial.size(); ++i) {
         assert(hm_table.count(partial[i]) == 1);
 
         int h = hm_table[partial[i]];
@@ -204,9 +204,9 @@ void HMHeuristic::generate_all_partial_tuple(tuple &base_tuple, tuple &t,
 
 
 int HMHeuristic::check_tuple_in_tuple(const tuple &tup, const tuple &big_tuple) {
-    for (int i = 0; i < tup.size(); ++i) {
+    for (size_t i = 0; i < tup.size(); ++i) {
         bool found = false;
-        for (int j = 0; j < big_tuple.size(); ++j) {
+        for (size_t j = 0; j < big_tuple.size(); ++j) {
             if (tup[i] == big_tuple[j]) {
                 found = true;
                 break;
