+++ src/search/landmarks/h_m_landmarks.cc
@@ -141,7 +141,7 @@ void HMLandmarks::get_m_sets_(int m, int num_included, int current_var,
     // include a value of current_var in the set
     for (int i = 0; i < g_variable_domain[current_var]; ++i) {
         bool use_var = true;
-        for (int j = 0; j < current.size(); ++j) {
+        for (size_t j = 0; j < current.size(); ++j) {
             if (!interesting(current_var, i, current[j].first, current[j].second)) {
                 use_var = false;
                 break;
@@ -176,7 +176,7 @@ void HMLandmarks::get_m_sets_of_set_(int m, int num_included, int current_var_in
     }
 
     bool use_var = true;
-    for (int i = 0; i < current.size(); ++i) {
+    for (size_t i = 0; i < current.size(); ++i) {
         if (!interesting(superset[current_var_index].first, superset[current_var_index].second,
                          current[i].first, current[i].second)) {
             use_var = false;
@@ -224,7 +224,7 @@ void HMLandmarks::get_split_m_sets_(
     if ((ss1_var_index != superset1.size()) &&
         ((ss2_var_index == superset2.size()) ||
          (superset1[ss1_var_index] < superset2[ss2_var_index]))) {
-        for (int i = 0; i < current.size(); ++i) {
+        for (size_t i = 0; i < current.size(); ++i) {
             if (!interesting(superset1[ss1_var_index].first, superset1[ss1_var_index].second,
                              current[i].first, current[i].second)) {
                 use_var = false;
@@ -246,7 +246,7 @@ void HMLandmarks::get_split_m_sets_(
                           ss1_var_index + 1, ss2_var_index,
                           current, subsets, superset1, superset2);
     } else {
-        for (int i = 0; i < current.size(); ++i) {
+        for (size_t i = 0; i < current.size(); ++i) {
             if (!interesting(superset2[ss2_var_index].first, superset2[ss2_var_index].second,
                              current[i].first, current[i].second)) {
                 use_var = false;
@@ -300,7 +300,7 @@ void HMLandmarks::get_m_sets(int m,
                              std::vector<FluentSet> &subsets,
                              const State &s) {
     FluentSet state_fluents;
-    for (int i = 0; i < g_variable_domain.size(); ++i) {
+    for (size_t i = 0; i < g_variable_domain.size(); ++i) {
         state_fluents.push_back(std::make_pair(i, s[i]));
     }
     get_m_sets(m, subsets, state_fluents);
@@ -331,12 +331,12 @@ void get_operator_postcondition(int op_index, FluentSet &post) {
     const std::vector<Effect> &effects = op.get_effects();
     std::vector<bool> has_effect_on_var(g_variable_domain.size(), false);
 
-    for (int i = 0; i < effects.size(); ++i) {
+    for (size_t i = 0; i < effects.size(); ++i) {
         post.push_back(make_pair(effects[i].var, effects[i].val));
         has_effect_on_var[effects[i].var] = true;
     }
 
-    for (int i = 0; i < preconditions.size(); ++i) {
+    for (size_t i = 0; i < preconditions.size(); ++i) {
         if (!has_effect_on_var[preconditions[i].var])
             post.push_back(make_pair(preconditions[i].var, preconditions[i].val));
     }
@@ -353,16 +353,16 @@ void HMLandmarks::print_pm_op(const PMOp &op) {
     std::vector<int>::const_iterator v_it;
 
     for (v_it = op.pc.begin(); v_it != op.pc.end(); ++v_it) {
-        for (int j = 0; j < h_m_table_[*v_it].fluents.size(); ++j) {
+        for (size_t j = 0; j < h_m_table_[*v_it].fluents.size(); ++j) {
             pcs.insert(h_m_table_[*v_it].fluents[j]);
         }
     }
     for (v_it = op.eff.begin(); v_it != op.eff.end(); ++v_it) {
-        for (int j = 0; j < h_m_table_[*v_it].fluents.size(); ++j) {
+        for (size_t j = 0; j < h_m_table_[*v_it].fluents.size(); ++j) {
             effs.insert(h_m_table_[*v_it].fluents[j]);
         }
     }
-    for (int i = 0; i < op.cond_noops.size(); ++i) {
+    for (size_t i = 0; i < op.cond_noops.size(); ++i) {
         cond_pc.clear();
         cond_eff.clear();
         int pm_fluent, j;
@@ -371,7 +371,7 @@ void HMLandmarks::print_pm_op(const PMOp &op) {
             print_fluentset(h_m_table_[pm_fluent].fluents);
             std::cout << std::endl;
 
-            for (int k = 0; k < h_m_table_[pm_fluent].fluents.size(); ++k) {
+            for (size_t k = 0; k < h_m_table_[pm_fluent].fluents.size(); ++k) {
                 cond_pc.insert(h_m_table_[pm_fluent].fluents[k]);
             }
         }
@@ -386,7 +386,7 @@ void HMLandmarks::print_pm_op(const PMOp &op) {
             print_fluentset(h_m_table_[pm_fluent].fluents);
             std::cout << std::endl;
 
-            for (int k = 0; k < h_m_table_[pm_fluent].fluents.size(); ++k) {
+            for (size_t k = 0; k < h_m_table_[pm_fluent].fluents.size(); ++k) {
                 cond_eff.insert(h_m_table_[pm_fluent].fluents[k]);
             }
         }
@@ -407,7 +407,7 @@ void HMLandmarks::print_pm_op(const PMOp &op) {
         std::cout << " ";
     }
     std::cout << std::endl << "Conditionals: " << std::endl;
-    for (int i = 0; i < conds.size(); ++i) {
+    for (size_t i = 0; i < conds.size(); ++i) {
         std::cout << "Cond PC #" << i << ":" << std::endl << "\t";
         for (it = conds[i].first.begin(); it != conds[i].first.end(); ++it) {
             print_proposition(*it);
@@ -474,7 +474,7 @@ void HMLandmarks::build_pm_ops() {
 
     // transfer ops from original problem
     // represent noops as "conditional" effects
-    for (int i = 0; i < g_operators.size(); ++i) {
+    for (size_t i = 0; i < g_operators.size(); ++i) {
         PMOp &op = pm_ops_[i];
         op.index = op_count++;
 
@@ -492,7 +492,7 @@ void HMLandmarks::build_pm_ops() {
         // set unsatisfied pc count for op
         unsat_pc_count_[i].first = pc_subsets.size();
 
-        for (int j = 0; j < pc_subsets.size(); ++j) {
+        for (size_t j = 0; j < pc_subsets.size(); ++j) {
             assert(set_indices_.find(pc_subsets[j]) != set_indices_.end());
             set_index = set_indices_[pc_subsets[j]];
             op.pc.push_back(set_index);
@@ -504,7 +504,7 @@ void HMLandmarks::build_pm_ops() {
         get_m_sets(m_, eff_subsets, eff);
         op.eff.reserve(eff_subsets.size());
 
-        for (int j = 0; j < eff_subsets.size(); ++j) {
+        for (size_t j = 0; j < eff_subsets.size(); ++j) {
             assert(set_indices_.find(eff_subsets[j]) != set_indices_.end());
             set_index = set_indices_[eff_subsets[j]];
             op.eff.push_back(set_index);
@@ -538,7 +538,7 @@ void HMLandmarks::build_pm_ops() {
                 unsat_pc_count_[i].second.push_back(noop_pc_subsets.size());
 
                 // push back all noop preconditions
-                for (int j = 0; j < noop_pc_subsets.size(); ++j) {
+                for (size_t j = 0; j < noop_pc_subsets.size(); ++j) {
                     assert(noop_pc_subsets[j].size() <= m_);
                     assert(set_indices_.find(noop_pc_subsets[j]) != set_indices_.end());
 
@@ -552,7 +552,7 @@ void HMLandmarks::build_pm_ops() {
                 this_cond_noop.push_back(-1);
 
                 // and the noop effects
-                for (int j = 0; j < noop_eff_subsets.size(); ++j) {
+                for (size_t j = 0; j < noop_eff_subsets.size(); ++j) {
                     assert(noop_eff_subsets[j].size() <= m_);
                     assert(set_indices_.find(noop_eff_subsets[j]) != set_indices_.end());
 
@@ -596,7 +596,7 @@ void HMLandmarks::init() {
     //  std::cout << "P^m index\tP fluents" << std::endl;
 
     // map each set to an integer
-    for (int i = 0; i < msets.size(); ++i) {
+    for (size_t i = 0; i < msets.size(); ++i) {
         h_m_table_.push_back(HMEntry());
         set_indices_[msets[i]] = i;
         h_m_table_[i].fluents = msets[i];
@@ -626,7 +626,7 @@ void HMLandmarks::calc_achievers() {
 
         std::set<int> candidates;
         // put all possible adders in candidates set
-        for (int i = 0; i < lmn.vars.size(); ++i) {
+        for (size_t i = 0; i < lmn.vars.size(); ++i) {
             const std::vector<int> &ops =
                 lm_graph->get_operators_including_eff(std::make_pair(lmn.vars[i],
                                                                      lmn.vals[i]));
@@ -690,7 +690,7 @@ void HMLandmarks::free_unneeded_memory() {
 void HMLandmarks::propagate_pm_fact(int factindex, bool newly_discovered,
                                     TriggerSet &trigger) {
     // for each action/noop for which fact is a pc
-    for (int i = 0; i < h_m_table_[factindex].pc_for.size(); ++i) {
+    for (size_t i = 0; i < h_m_table_[factindex].pc_for.size(); ++i) {
         std::pair<int, int> const &info = h_m_table_[factindex].pc_for[i];
 
         // a pc for the action itself
@@ -731,7 +731,7 @@ void HMLandmarks::compute_h_m_landmarks() {
     TriggerSet current_trigger, next_trigger;
 
     // for all of the initial state <= m subsets, mark level = 0
-    for (int i = 0; i < init_subsets.size(); ++i) {
+    for (size_t i = 0; i < init_subsets.size(); ++i) {
         int index = set_indices_[init_subsets[i]];
         h_m_table_[index].level = 0;
 
@@ -740,7 +740,7 @@ void HMLandmarks::compute_h_m_landmarks() {
     }
 
     // mark actions with no precondition to be applied
-    for (int i = 0; i < pm_ops_.size(); ++i) {
+    for (size_t i = 0; i < pm_ops_.size(); ++i) {
         if (unsat_pc_count_[i].first == 0) {
             // create empty set or clear prev entries
             current_trigger[i].clear();
@@ -810,7 +810,7 @@ void HMLandmarks::compute_h_m_landmarks() {
             // landmarks changed for action itself, have to recompute
             // landmarks for all noop effects
             if (op_it->second.empty()) {
-                for (int i = 0; i < action.cond_noops.size(); ++i) {
+                for (size_t i = 0; i < action.cond_noops.size(); ++i) {
                     // actions pcs are satisfied, but cond. effects may still have
                     // unsatisfied pcs
                     if (unsat_pc_count_[op_index].second[i] == 0) {
@@ -941,7 +941,7 @@ void HMLandmarks::generate_landmarks() {
     std::vector<FluentSet> goal_subsets;
     get_m_sets(m_, goal_subsets, g_goal);
     std::list<int> all_lms;
-    for (int i = 0; i < goal_subsets.size(); ++i) {
+    for (size_t i = 0; i < goal_subsets.size(); ++i) {
         /*
            std::cout << "Goal set: ";
            print_fluentset(goal_subsets[i]);
